Update integration tests to expected new output
