+++ sqldelight-compiler/integration-tests/src/test/kotlin/com/example/GroupQueries.kt
@@ -1,9 +1,15 @@
 package com.example
 
 import app.cash.sqldelight.Query
-import app.cash.sqldelight.Transacter
+import app.cash.sqldelight.TransacterImpl
+import app.cash.sqldelight.db.SqlDriver
 import kotlin.Long
 
-public interface GroupQueries : Transacter {
-  public fun selectAll(): Query<Long>
+public class GroupQueries(
+  private val driver: SqlDriver
+) : TransacterImpl(driver) {
+  public fun selectAll(): Query<Long> = Query(165688501, arrayOf("group"), driver, "Group.sq",
+      "selectAll", "SELECT `index` FROM `group`") { cursor ->
+    cursor.getLong(0)!!
+  }
 }
